Structure the workflow of the HTML generator.
It will go from package level to revision to target.
